requirements.txt
tensorflow>=2.12.0
tensorflow-hub>=0.13.0
numpy>=1.21.0
scikit-learn>=1.0.2
Pillow>=9.0.0
tqdm>=4.65.0
opencv-python>=4.5.1
pillow-heif>=0.12.0
h5py>=3.1.0